Deadline May 3, 2020
Write a Java program equivalent to the Python program at:

https://github.com/ptarau/CrackerBarrelPegPuzzle

that implements a solver for the Cracker Barrel Peg Puzzle at:

https://shop.crackerbarrel.com/toys-games/games/travel-games/peg-game/606154

This time you will need to use classes and objects for things like the moves, the board you display them on and the solver itself.

Post your answer on github.com and provide a link to it in canvas. You can reuse the repository created for the same assignment in Prolog or create a new one.